backend/requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
g2pk2
jamo
